separate out processor module and make isomorphic files
