+++ my-first-workflow/chat_1/markdown_body.md
@@ -1,0 +1,1 @@
+**{{ recipient.name }}** performed an action.
