+++ my-first-workflow/email_1/visual_blocks/1.content.md
@@ -1,0 +1,7 @@
+{% for activity in activities %}
+<p>{{ activity.actor.name }} commented</p>
+{% endfor %}
+
+hey
+
+asdf
