+++ my-first-workflow/email_2/visual_blocks/1.content.md
@@ -1,0 +1,1 @@
+Hey **{{ recipient.name | split: " " | first }}** 👋 - {{ actor.name }} completed an activity.
