+++ my-first-workflow/in_app_feed_1/markdown_body.md
@@ -1,0 +1,1 @@
+Hey **{{ recipient.name | split: ' ' | first }}** 👋 - {{ actor.name }} completed an activity.
