+++ my-first-workflow/sms_1/text_body.txt
@@ -1,0 +1,3 @@
+Hey {{ recipient.name | split: ' ' | first }} - {{ actor.name }} completed an activity.
+
+hey
