+++ src/lib/marshal/workflow/index.ts
@@ -3,3 +3,4 @@ export * from "./helpers";
 export * from "./reader";
 export * from "./types";
 export * from "./writer";
+export * from "./processor.isomorphic";
